+++ cmd/kobs/satellite/satellite.go
@@ -11,6 +11,7 @@ import (
 	"github.com/kobsio/kobs/pkg/metrics"
 	"github.com/kobsio/kobs/pkg/satellite"
 	"github.com/kobsio/kobs/pkg/satellite/plugins"
+	"github.com/kobsio/kobs/pkg/tracer"
 	"github.com/kobsio/kobs/pkg/version"
 
 	"github.com/spf13/cobra"
@@ -42,6 +43,18 @@ var Cmd = &cobra.Command{
 
 		log.Info(nil, "Version information", version.Info()...)
 		log.Info(nil, "Build context", version.BuildContext()...)
+
+		traceEnabled, _ := cmd.Flags().GetBool("trace.enabled")
+		traceServiceName, _ := cmd.Flags().GetString("trace.service-name")
+		traceProvider, _ := cmd.Flags().GetString("trace.provider")
+		traceAddress, _ := cmd.Flags().GetString("trace.address")
+
+		if traceEnabled {
+			err := tracer.Setup(traceServiceName, traceProvider, traceAddress)
+			if err != nil {
+				log.Fatal(nil, "Could not setup tracing", zap.Error(err), zap.String("provider", traceProvider), zap.String("address", traceAddress))
+			}
+		}
 
 		// Load the configuration for the satellite from the provided configuration file.
 		cfg, err := config.Load(satelliteConfigFile)