+++ pkg/hub/hub.go
@@ -19,6 +19,7 @@ import (
 	"github.com/kobsio/kobs/pkg/log"
 	"github.com/kobsio/kobs/pkg/middleware/debug"
 	"github.com/kobsio/kobs/pkg/middleware/httplog"
+	"github.com/kobsio/kobs/pkg/middleware/httptracer"
 	"github.com/kobsio/kobs/pkg/middleware/metrics"
 
 	"github.com/go-chi/chi/v5"
@@ -89,6 +90,7 @@ func New(config api.Config, debugUsername, debugPassword, hubAddress string, aut
 		r.Use(middleware.Recoverer)
 		r.Use(middleware.URLFormat)
 		r.Use(userauth.Handler(authEnabled, authHeaderUser, authHeaderTeams, authSessionToken, authSessionInterval, storeClient))
+		r.Use(httptracer.Handler("hub"))
 		r.Use(metrics.Metrics)
 		r.Use(httplog.Logger)
 		r.Use(render.SetContentType(render.ContentTypeJSON))