+++ pkg/satellite/api/users/users_test.go
@@ -44,7 +44,7 @@ func TestGetUsers(t *testing.T) {
 
 			mockClustersClient := &clusters.MockClient{}
 			mockClustersClient.AssertExpectations(t)
-			mockClustersClient.On("GetClusters").Return([]cluster.Client{mockClusterClient})
+			mockClustersClient.On("GetClusters", mock.Anything).Return([]cluster.Client{mockClusterClient})
 
 			tt.prepare(mockClusterClient)