+++ plugins/plugin-flux/pkg/sync/sync.go
@@ -35,7 +35,7 @@ func doReconcileAnnotations(annotations map[string]string) {
 // Kustomization can be used to sync a Flux Kustomization. For that the cluster, namespace and name for the resource
 // must be provided.
 func Kustomization(ctx context.Context, clustersClient cluster.Client, namespace, name string) error {
-	client, err := clustersClient.GetClient(createScheme())
+	client, err := clustersClient.GetClient(ctx, createScheme())
 	if err != nil {
 		return fmt.Errorf("could not get client: %w", err)
 	}
@@ -62,7 +62,7 @@ func Kustomization(ctx context.Context, clustersClient cluster.Client, namespace
 // HelmRelease can be used to sync a Flux HelmRelease. For that the cluster, namespace and name for the resource must be
 // provided.
 func HelmRelease(ctx context.Context, clustersClient cluster.Client, namespace, name string) error {
-	client, err := clustersClient.GetClient(createScheme())
+	client, err := clustersClient.GetClient(ctx, createScheme())
 	if err != nil {
 		return fmt.Errorf("could not get client: %w", err)
 	}