+++ plugins/plugin-harbor/pkg/instance/instance.go
@@ -7,7 +7,9 @@ import (
 	"net/url"
 
 	"github.com/kobsio/kobs/pkg/middleware/roundtripper"
+
 	"github.com/mitchellh/mapstructure"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 )
 
 // Config is the structure of the configuration for a single Harbor database instance.
@@ -152,7 +154,7 @@ func New(name string, options map[string]interface{}) (Instance, error) {
 		name:    name,
 		address: config.Address,
 		client: &http.Client{
-			Transport: roundTripper,
+			Transport: otelhttp.NewTransport(roundTripper),
 		},
 	}, nil
 }