+++ plugins/plugin-helm/cmd/helm.go
@@ -40,7 +40,7 @@ func appendIfMissing(items []string, item string) []string {
 }
 
 func (router *Router) getClusters(w http.ResponseWriter, r *http.Request) {
-	clusters := router.clustersClient.GetClusters()
+	clusters := router.clustersClient.GetClusters(r.Context())
 
 	var clusterNames []string
 	for _, cluster := range clusters {
@@ -55,7 +55,7 @@ func (router *Router) getNamespaces(w http.ResponseWriter, r *http.Request) {
 	clusterNames := r.URL.Query()["cluster"]
 
 	for _, clusterName := range clusterNames {
-		tmpNamespaces, err := router.clustersClient.GetCluster(clusterName).GetNamespaces(r.Context())
+		tmpNamespaces, err := router.clustersClient.GetCluster(r.Context(), clusterName).GetNamespaces(r.Context())
 		if err != nil {
 			log.Error(r.Context(), "Could not get namespaces", zap.Error(err), zap.String("cluster", clusterName))
 			errresponse.Render(w, r, err, http.StatusBadRequest, "Could not get namespaces")
@@ -88,7 +88,7 @@ func (router *Router) getReleases(w http.ResponseWriter, r *http.Request) {
 	var helmReleases []*client.Release
 
 	for _, clusterName := range clusterNames {
-		cluster := router.clustersClient.GetCluster(clusterName)
+		cluster := router.clustersClient.GetCluster(r.Context(), clusterName)
 		if cluster == nil {
 			log.Error(r.Context(), "Invalid cluster name", zap.String("cluster", clusterName))
 			errresponse.Render(w, r, nil, http.StatusBadRequest, "Invalid cluster name")
@@ -162,7 +162,7 @@ func (router *Router) getRelease(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	cluster := router.clustersClient.GetCluster(clusterName)
+	cluster := router.clustersClient.GetCluster(r.Context(), clusterName)
 	if cluster == nil {
 		log.Error(r.Context(), "Invalid cluster name", zap.String("cluster", clusterName))
 		errresponse.Render(w, r, nil, http.StatusBadRequest, "Invalid cluster name")
@@ -201,7 +201,7 @@ func (router *Router) getReleaseHistory(w http.ResponseWriter, r *http.Request) 
 		return
 	}
 
-	cluster := router.clustersClient.GetCluster(clusterName)
+	cluster := router.clustersClient.GetCluster(r.Context(), clusterName)
 	if cluster == nil {
 		log.Error(r.Context(), "Invalid cluster name", zap.String("cluster", clusterName))
 		errresponse.Render(w, r, nil, http.StatusBadRequest, "Invalid cluster name")